Fixed problem with generate_sonic_hash, where averages weren't correctly calculated. Also changed return type.

+++ libwaveplot/waveplot.h
@@ -21,6 +21,7 @@
 #define __LIBWP_WAVEPLOT_H__
 
 #include <stdlib.h>
+#include <stdint.h>
 
 #include "dllexport.h"
 #include "typedefs.h"
@@ -49,6 +50,6 @@ LIBWAVEPLOT_API void update_waveplot(waveplot_t* waveplot, audio_samples_t* samp
 LIBWAVEPLOT_API void finish_waveplot(waveplot_t* waveplot);
 LIBWAVEPLOT_API void resample_waveplot(waveplot_t* waveplot, size_t target_length, size_t target_amplitude);
 
-LIBWAVEPLOT_API unsigned int generate_sonic_hash(waveplot_t* waveplot);
+LIBWAVEPLOT_API uint16_t generate_sonic_hash(waveplot_t* waveplot);
 
 #endif //__LIBWP_WAVEPLOT_H__
